# Optimized conversations existence