--- a/lib/loaders/api/loader_with_authentication_factory.js
+++ b/lib/loaders/api/loader_with_authentication_factory.js
@@ -19,11 +19,12 @@ import { verbose, error } from "lib/loggers"
  */
 export const apiLoaderWithAuthenticationFactory = api => {
   return accessTokenLoader => {
-    return (path, globalParams = {}, apiOptions = {}) => {
+    return (path, globalParams = {}, apiOptions = {}, optimizeRequests = null) => {
       const loader = new DataLoader(
-        keys => {
-          return accessTokenLoader().then(accessToken =>
-            Promise.all(
+        allKeys => {
+          const { keys, postProcess } = optimizeRequests ? optimizeRequests(allKeys) : { keys: allKeys }
+          return accessTokenLoader().then(accessToken => {
+            const responsesPromise = Promise.all(
               keys.map(key => {
                 const clock = timer(key)
                 clock.start()
@@ -45,7 +46,8 @@ export const apiLoaderWithAuthenticationFactory = api => {
                 })
               })
             )
-          )
+            return postProcess ? responsesPromise.then(responses => postProcess(responses)) : responsesPromise
+          })
         },
         {
           batch: true,