--- a/lib/loaders/api/loader_without_authentication_factory.js
+++ b/lib/loaders/api/loader_without_authentication_factory.js
@@ -22,10 +22,11 @@ import { verbose, error } from "lib/loggers"
  * @param {apiSignature} api a function that performs an API request
  */
 export const apiLoaderWithoutAuthenticationFactory = api => {
-  return (path, globalParams = {}, apiOptions = {}) => {
+  return (path, globalParams = {}, apiOptions = {}, optimizeRequests = null) => {
     const loader = new DataLoader(
-      keys =>
-        Promise.all(
+      allKeys => {
+        const { keys, postProcess } = optimizeRequests ? optimizeRequests(allKeys) : { keys: allKeys }
+        const responsesPromise = Promise.all(
           keys.map(key => {
             const clock = timer(key)
             clock.start()
@@ -71,9 +72,11 @@ export const apiLoaderWithoutAuthenticationFactory = api => {
               )
             })
           })
-        ),
+        )
+        return postProcess ? responsesPromise.then(responses => postProcess(responses)) : responsesPromise
+      },
       {
-        batch: false,
+        batch: !!optimizeRequests, // FIXME Does this ever need to be false? I guess requests will immediately resolve.
         cache: true,
       }
     )