With these changes a ‘count’ request will be merged with other data requests, if they exist, but only ever 1 request will be made to fulfil both these queries.

## Just count (in form of boolean)

```graphql
{
  me {
    has_conversations
  }
}
```

```
Elapsed: 0s 576.05448ms - conversations?from_id=[SNIP]&from_type=User&size=0
```

## With conversation data

```graphql
{
  me {
    has_conversations
    conversations(first: 10) {
      totalCount
      edges {
        node {
          initial_message
        }
      }
    }
  }
}
```

```
Elapsed: 0s 983.291054ms - conversations?from_id=[SNIP]&from_type=User&page=1&size=10
```