@alloy 2017-09-29
Note to readers: exposing the `has_conversations` field and optimising the requests is mostly to figure out the possibilities in request optimisations, it may not be the perfect candidate for optimisations in general (the extra count request would probably have been fine).