@mzikherman 2017-09-29
Looks great!

My honest concern/thought is that it's a lot of work/code, machinations and prying into internals, (and possibly even brittle/buggy based on how loading conversations may change in the future), and I'm just not sure if optimizing the count loaders is really necessary?

These count queries are some of the lightest weight queries we can make. It just feels like a premature optimization? I'd like to see some proof/evidence over what this will save, in order for it to be worth it, IMO.