@alloy 2017-09-29
> These count queries are some of the lightest weight queries we can make

I think you’re only saying this from a backend perspective? In this case I care less about that (because they are indeed probably cheap), but more about the client’s perspective and the possibilty for an additional 0.5 sec of request time.

In short, I think that going forward we should take a few minutes to make all (new) requests more optimal than just completely naive. However, the reality is that that’s probably not going to happen if this foundation is not in place yet and there’s no example loaders/resolvers to learn from.

> machinations and prying into internals, (and possibly even brittle/buggy based on how loading conversations may change in the future)

I agree. I’d like to immediately add typings once @orta’s PR that enables TS type-checking is merged so that this should become less of a concern.