@mzikherman 2017-09-29
I definitely can't argue with the learning experimental aspect of this, I'm just wondering as to the production-worthiness.

I'm a bit unclear as to how front-end clients would be affected. If keys are being loaded and resolved in parallel, might the case of `has_conversations` along with an actual `conversations` query (the case we are trying to optimize), still result in just waiting for the longest request (likely the actual `conversations` load)? If so, then the savings isn't directly additive, and thus perhaps not as big an issue as we think?