@alloy 2017-09-29
> I'm just wondering as to the production-worthiness.

That’s fair, although I would counter that this part of the schema is all still to be extensively tested when we properly QA the app.

> I'm a bit unclear as to how front-end clients would be affected. If keys are being loaded and resolved in parallel, might the case of `has_conversations` along with an actual `conversations` query (the case we are trying to optimize), still result in just waiting for the longest request (likely the actual conversations load)?

Yes. But the point I’m trying to make is that this addition is more about having the foundation in and a contrived example of how to use the `optimizeRequest` parameter for cases where it would matter much more.