@alloy 2017-09-29
Btw on the point of production-worthiness, I should definitely add some test coverage.