@alloy 2017-10-13
Gonna close this for now, but leaving the branch as an example for the time being.